feat(claim.progress): prevent deleting progress step after 5 minutes of its creation

+++ src/modules/claim/progress/progress.controller.ts
@@ -3,6 +3,7 @@ import {
     Body,
     Controller,
     Delete,
+    ForbiddenException,
     HttpCode,
     NotFoundException,
     Param,
@@ -22,7 +23,7 @@ import { ClaimService } from '../claim.service';
 import { Languages } from '../../language/enums/languages.enums';
 import { isLanguage } from '../../../utils/isLanguage';
 import { CreateProgressDto } from './dto/create-progress.dto';
-import { INVALID_CLAIM_ID } from '../constants';
+import { INVALID_CLAIM_ID, MINUTE } from '../constants';
 import { ProgressVariants } from './constants/progresses/progressVariants';
 import { ClaimStatus } from '@prisma/client';
 import { InjectQueue } from '@nestjs/bullmq';
@@ -117,6 +118,12 @@ export class ProgressController {
 
         if (!progress) {
             throw new NotFoundException(INVALID_PROGRESS_ID);
+        }
+
+        if (Date.now() - progress.createdAt.getTime() > 1000 * MINUTE * 5) {
+            throw new ForbiddenException(
+                'You cannot delete a progress step after 5 minutes of its creation',
+            );
         }
 
         await this.progressesService.deleteProgress(progressId);
